backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.10.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
geopandas==1.0.1
shapely==2.0.6
networkx==3.4.2
momepy==0.9.0
fiona==1.10.1
pyproj==3.7.0
google-genai>=1.0.0
twilio==9.3.7
python-multipart==0.0.19
aiofiles==24.1.0
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
email-validator==2.2.0
mongomock-motor>=0.0.21
